% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{clean_storm_data}
\alias{clean_storm_data}
\title{Clean storm dataset}
\usage{
clean_storm_data(storm_data, include_narratives)
}
\arguments{
\item{storm_data}{A dataset of storm data. This dataset must include certain columns given
in the NOAA Storm Events datasets for which this package was created.}

\item{include_narratives}{A logical value for whether the final data
data frame should include columns for episode and event narratives
(TRUE) or not (FALSE, the default)}
}
\value{
A cleaned version of the dataset input to the function.
}
\description{
Cleans the storm dataset to prepare for further processing. This includes changing all
variable names to lowercase, removing some unneeded columns, and removing the narratives
if requested by the user.
}
