% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_events.R
\name{convert_damage_costs}
\alias{convert_damage_costs}
\title{Convert labels for damage costs to a prettier format}
\usage{
convert_damage_costs(x)
}
\arguments{
\item{x}{The levels of a factor, as formatted by \code{cut} for crop or
property damage values.}
}
\description{
Convert labels for damage costs to a prettier format
}
\examples{
crop_dmg_levels <- c("[0,20]", "(20,600]", "(600,1e+04]", "(1e+04,3e+05]")
convert_damage_costs(crop_dmg_levels)

}
