% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.fips}
\alias{county.fips}
\title{County FIPS Codes for United States}
\format{
## `county.fips`
A data frame with 3,117 rows and 3 columns:
\describe{
  \item{fips}{Number object with 5 Number FIPs Code for County}
  \item{state}{Character string of state county is located in}
  \item{county}{Character string with County name}
}
}
\usage{
county.fips
}
\description{
A dataframe of county fips codes for each state within the Unites States.
}
\examples{
data("county.fips")
head(county.fips)

}
\keyword{datasets}
