% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_damagecrops.R
\name{find_damage_crops}
\alias{find_damage_crops}
\title{Find all damaged crop listings for date range}
\usage{
find_damage_crops(date_range = NULL, ts_only = FALSE, dist_limit = NULL,
  storm = NULL)
}
\arguments{
\item{ts_only}{A logical value indicating whether to filter events to only
those in tropical storm-related categories.}

\item{begin_date}{A character string giving the date, in the format
"\%Y-\%m-\%d".}

\item{end_date}{A character string giving the date, in the format
"\%Y-\%m-\%d". The end date must be in the same year as \code{begin_date}.}
}
\description{
This function will find all of the crops damaged in the US for a specified date
range.
}
\examples{
find_damage_crops(date_range = c("1999-10-15", "1999-10-20"))

find_damage_crops(date_range = c("1999-10-16", "1999-10-18"),
   storm = "Floyd-1999", dist_limit = 200)

}

