% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_damagecrops.R, R/map_damageproperty.R, R/map_directdeaths.R, R/map_directinjuries.R, R/map_events.R, R/map_indirectdeaths.R, R/map_indirectinjuries.R
\name{find_file_name}
\alias{find_file_name}
\title{Find the file name for a specific year}
\usage{
find_file_name(year = NULL)

find_file_name(year = NULL)

find_file_name(year = NULL)

find_file_name(year = NULL)

find_file_name(year = NULL)

find_file_name(year = NULL)

find_file_name(year = NULL)
}
\arguments{
\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}

\item{year}{A character string giving the year.}
}
\description{
This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.

This function will find the name of the detailed file from Storm Events Database
for a specific year.
}
\examples{
find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

find_file_name(year = 1999)

find_file_name(year = 2003)

}

