% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_indirectinjuries.R
\name{find_indirect_injuries}
\alias{find_indirect_injuries}
\title{Find all indirect injury listings for date range}
\usage{
find_indirect_injuries(first_date = NULL, last_date = NULL,
  ts_only = FALSE, dist_limit = NULL, storm = NULL)
}
\arguments{
\item{ts_only}{A logical value indicating whether to filter events to only
those in tropical storm-related categories.}

\item{begin_date}{A character string giving the date, in the format
"%Y-%m-%d".}

\item{end_date}{A character string giving the date, in the format
"%Y-%m-%d". The end date must be in the same year as \code{begin_date}.}
}
\description{
This function will find all of the indirect injuries in the US for a specified date
range.
}
\examples{
find_indirect_injuries(first_date = "1999-10-15", last_date = "1999-10-20")

find_indirect_injuries(first_date = "1999-10-16", last_date = "1999-10-18",
   storm = "Floyd-1999", dist_limit = 200)

}

