% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_zone_data}
\alias{forecast_zone_data}
\title{FIPS Codes for Forecast Zones within United States}
\format{
## `forecast_zone_date`
A data frame with 4,762 rows and 4 columns:
\describe{
  \item{STATE}{Character string of Abbreviated State Name County is Located in}
  \item{NAME}{Character string with forecast zone name}
  \item{COUNTY}{Character string with County name}
  \item{FIPS}{Integer object with 5 Number FIPs Code for County}
}
}
\usage{
forecast_zone_data
}
\description{
A dataframe of county fips codes associated with forecast zones within the Unites States.
}
\examples{
data("forecast_zone_data")
head(forecast_zone_data)

}
\keyword{datasets}
