% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_events.R
\name{get_county_map}
\alias{get_county_map}
\title{Get map data for counties}
\usage{
get_county_map(states = "east")
}
\arguments{
\item{states}{A character string specifying either a state name or names or one of
"all" (map all states in the continental US) or "east" (plot states in the
Eastern half of the US. The default is "east".}
}
\value{
A dataframe with map data pulled using the \code{map_data} function in
   \code{ggplot2}, filtered to states in the eastern half of the United States
   if the user specifies \code{east_only}.
}
\description{
Get map data for counties
}
