% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_damagecrops.R
\name{map_damage_crops}
\alias{map_damage_crops}
\title{Map crops damaged for a date range}
\usage{
map_damage_crops(first_date = NULL, last_date = NULL, ts_only = FALSE,
  east_only = TRUE, dist_limit = NULL, storm = NULL, add_tracks = FALSE)
}
\arguments{
\item{ts_only}{A logical value indicating whether to filter events to only
those in tropical storm-related categories.}

\item{east_only}{A logical value specifying whether to restrict the map to
the eastern half of the United States (default is TRUE).}

\item{add_tracks}{A logical value specifying whether to add the tracks of
a hurricane to the map (default = FALSE).}
}
\description{
This function maps all crops damaged listed with a starting date within a
specified date range.
}
\examples{
map_damage_crops(first_date = "1999-10-15", last_date = "1999-10-20")
map_damage_crops(first_date = "1999-10-16", last_date = "1999-10-18",
   east_only = FALSE, ts_only = TRUE)
map_damage_crops(first_date = "1999-10-16", last_date = "1999-10-18")
map_damage_crops(first_date = "1999-10-16", last_date = "1999-10-18",
   dist_limit = 100, storm = "Floyd-1999",
    add_tracks = TRUE)

}

