% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_directdeaths.R
\name{map_direct_deaths}
\alias{map_direct_deaths}
\title{Map direct deaths for a date range}
\usage{
map_direct_deaths(date_range = NULL, event_types = NULL, east_only = TRUE,
  dist_limit = NULL, storm = NULL, add_tracks = FALSE)
}
\arguments{
\item{date_range}{A character vector of length two with the start and end
dates to pull data for (e.g., \code{c("1999-10-16", "1999-10-18")}).}

\item{event_types}{Character vector with the types of storm events that should be kept.
The default value (NULL) keeps all types of events. See the "Details" vignette for this
package for more details on possible event types.}

\item{dist_limit}{A numeric scalar with the distance (in kilometers) that a county
must be from the storm's path to be included. The default (NULL) does not eliminate any
events based on distance from a storm's path. This option should only be used when also
specifying a storm with the \code{storm} parameter.}

\item{storm}{A character string with the name of the storm to pull storm
events data for. This string must follow the format
"[storm-name]-[4-digit storm year]" (e.g., \code{"Floyd-1999"}).
Currently, this functionality only works for storms included in the
extended hurricane best tracks, which covers 1988 to 2015.}

\item{add_tracks}{A logical value specifying whether to add the tracks of
a hurricane to the map (default = FALSE).}
}
\description{
This function maps all direct deaths listed with a starting date within a
specified date range.
}
\examples{
\dontrun{
map_direct_deaths(date_range = c("1999-09-01", "1999-09-30"))
map_direct_deaths(date_range = c("1999-09-20", "1999-09-30"),
   east_only = FALSE, event_type = c("Flood","Flash Flood"))
map_direct_deaths(date_range = c("1999-09-01", "1999-09-30"))
map_direct_deaths(date_range = c("1999-09-01", "1999-09-30"),
   dist_limit = 100, storm = "Floyd-1999",
    add_tracks = TRUE)
map_direct_deaths(storm = "Floyd-1999", dist_limit = 250, add_tracks = TRUE)
}

}
