% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_events.R
\name{map_events}
\alias{map_events}
\title{Map storm events for a date range}
\usage{
map_events(first_date, last_date, ts_only = FALSE, east_only = TRUE,
  plot_type = "any events", dist_limit = NULL, storm = NULL,
  add_tracks = FALSE)
}
\arguments{
\item{ts_only}{A logical value indicating whether to filter events to only
those in tropical storm-related categories.}

\item{east_only}{A logical value specifying whether to restrict the map to
the eastern half of the United States (default is TRUE).}

\item{add_tracks}{A logical value specifying whether to add the tracks of
a hurricane to the map (default = FALSE).}
}
\description{
This function maps all storm events listed with a starting date within a
specified date range.
}
\examples{
map_events(first_date = "1999-10-15", last_date = "1999-10-20")
map_events(first_date = "1999-10-16", last_date = "1999-10-18",
   east_only = FALSE, ts_only = TRUE)
map_events(first_date = "1999-10-16", last_date = "1999-10-18",
   plot_type = "number of events")
map_events(first_date = "1999-10-16", last_date = "1999-10-18",
   dist_limit = 100, storm = "Floyd-1999",
    add_tracks = TRUE, plot_type = "number of events")

}

