% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_indirectinjuries.R
\name{map_indirect_injuries}
\alias{map_indirect_injuries}
\title{Map indirect injuries for a date range}
\usage{
map_indirect_injuries(date_range = NULL, event_type = NULL,
  east_only = TRUE, dist_limit = NULL, storm = NULL, add_tracks = FALSE)
}
\arguments{
\item{date_range}{A character vector of length two with the start and end
dates to pull data for (e.g., \code{c("1999-10-16", "1999-10-18")}).}

\item{event_type}{Specifies the types of storm events should be included.}

\item{east_only}{A logical value specifying whether to restrict the map to
the eastern half of the United States (default is TRUE).}

\item{dist_limit}{If selected, the distance (in kilometers) that a county
must be from the storm's path to be included.}

\item{storm}{A character string with the name of the storm to pull storm
events data for. This string must follow the format
"[storm-name]-[4-digit storm year]" (e.g., \code{"Floyd-1999"}).
Currently, this functionality only works for storms included in the
extended hurricane best tracks, which covers 1988 to 2015.}

\item{add_tracks}{A logical value specifying whether to add the tracks of
a hurricane to the map (default = FALSE).}
}
\description{
This function maps all indirect injuries listed with a starting date within a
specified date range.
}
\examples{
\dontrun{
map_indirect_injuries(date_range = c("1999-09-10", "1999-09-30"))
map_indirect_injuries(date_range = c("1999-09-01", "1999-09-30"),
                      east_only = FALSE, event_type = c("Flood","Flash Flood"))
map_indirect_injuries(date_range = c("1999-09-10", "1999-09-30"))
map_indirect_injuries(dist_limit = 100, storm = "Floyd-1999",
                      add_tracks = TRUE)
}

}
