% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_input_args}
\alias{process_input_args}
\title{Process inputs to main functions}
\usage{
process_input_args(date_range = NULL, storm = NULL)
}
\value{
A list with date ranges and storm identification based on
   user inputs to arguments in a main package function.
}
\description{
Processes some of the user's inputs for arguments for main
package functions, looks for any errors in input, and determines
elements like the year or years of storm data needed based on
user inputs.
}
