% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_events.R
\name{find_events}
\alias{find_events}
\title{Find all event listings for date range}
\usage{
find_events(date_range = NULL, event_type = NULL, dist_limit = NULL,
  storm = NULL, include_narratives = FALSE, include_ids = FALSE)
}
\arguments{
\item{date_range}{A character vector of length two with the start and end
dates to pull data for (e.g., \code{c("1999-10-16", "1999-10-18")}).}

\item{event_type}{Specifies the types of storm events should be included.}

\item{dist_limit}{If selected, the distance (in kilometers) that a county
must be from the storm's path to be included.}

\item{storm}{A character string with the name of the storm to pull storm
events data for. This string must follow the format
"[storm-name]-[4-digit storm year]" (e.g., \code{"Floyd-1999"}).
Currently, this functionality only works for storms included in the
extended hurricane best tracks, which covers 1988 to 2014.}

\item{include_narratives}{A logical value for whether the final data
data frame should include columns for episode and event narratives
(TRUE) or not (FALSE, the default)}

\item{include_ids}{A logical value for whether the final data frame
should include columns for event and episode IDs (TRUE) or not
(FALSE, the default). If included, these IDs could be used in some
cases to link events to data in the "fatalities" or "locations"
files available through the NOAA Storm Events database.}
}
\description{
This function will find all of the events in the US for a specified date
range.
}
\examples{
\dontrun{
# Events by date range
find_events(date_range = c("1999-09-10", "1999-09-30"))

# Events within a certain distance and time range of a tropical storm
find_events(storm = "Floyd-1999", dist_limit = 200)
}

}

