% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{find_file_name}
\alias{find_file_name}
\title{Find the file name for a specific year}
\usage{
find_file_name(year = NULL, file_type = "detail")
}
\arguments{
\item{year}{A character string giving the year.}

\item{file_type}{The type of file you would like to pull. Choices include:
"details" (the default), "fatalities", or "locations".}
}
\description{
This function will find the name of the detailed file from Storm Events Database
for a specific year.
}
\examples{
find_file_name(year = 1999)
find_file_name(year = 2003, file_type = "fatalities")

}

