% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{match_forecast_county}
\alias{match_forecast_county}
\title{Match events by forecast zone to county}
\usage{
match_forecast_county(storm_data_z)
}
\arguments{
\item{storm_data_z}{A dataframe of storm events reported by forecast zone
(i.e., \code{cz_type == "Z"}) rather than county. This dataframe should
include the columns:
\begin{itemize}
  \item{\code{state}:}{State name, in lowercase}
  \item{\code{cz_name}:}{Location name, in lowercase}
  \item{\code{cz_fips}:}{Forecast zone FIPS}
\end{itemize}}
}
\value{
The dataframe of events input to the function, with county FIPS
   added for events matched to a county in the \code{fips} column. Events
   that could not be matched are kept in the dataframe, but the \code{fips}
   code is set to \code{NA}.
}
\description{
For events reported by forecast zone, use regular expressions to match
as many as possible to counties.
}
\details{
This function tries to match the \code{cz_name} of each event to
   a state and county name from the \code{county.fips} dataframe that comes
   with the \code{maps} package.
}
\note{
This function does not provide any matches for events outside
   of the continental U.S.
}
