% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{match_forecast_county}
\alias{match_forecast_county}
\title{Match events by forecast zone to county, OLD Function}
\usage{
match_forecast_county(storm_data_z)
}
\arguments{
\item{storm_data_z}{A dataframe of storm events reported by forecast zone
(i.e., \code{cz_type == "Z"}) rather than county. This dataframe should
include the columns:
\itemize{
  \item{\code{state}: }{State name, in lowercase}
  \item{\code{cz_name}: }{Location name, in lowercase}
  \item{\code{cz_fips}: }{Forecast zone FIPS}
}}
}
\value{
The dataframe of events input to the function, with county FIPS
   added for events matched to a county in the \code{fips} column. Events
   that could not be matched are kept in the dataframe, but the \code{fips}
   code is set to \code{NA}.
}
\description{
For events reported by forecast zone, use regular expressions to match
as many as possible to counties.
}
\details{
This function tries to join storm event data with FIPS data provided by
   the NOAA Forecast Zone Data set. The following steps are taken to join \code{eventid}
   to at least one \code{FIPS}:
   \enumerate{
    "cz_fips" = "ZONE"
     \item Tries to match \code{cz_fips} to \code{ZONE} in \code{forecast_zone_data}.
     \item Next, match \code{cz_name} to \code{cz_name} in \code{forecast_zone_data}.
   }
}
\note{
This function does not provide any matches for events outside
   of the continental U.S.

An older version of this function is provided with \code{old_match_forecast_county} that uses
   string patterns to join county FIPS codes. This old version may be better for older storm event data.
   It is recommended to try both and see which version produces better results.
}
\examples{
counties_to_parse <- dplyr::tibble(event_id = c(1:18),
                                  cz_name = c("UNCOMPAHGRE PLATEAU AND DALLAS DIVIDE",
                                              "NORTH SNOWY RANGE FOOTHILLS",
                                              "CRAZY MOUNTAINS",
                                              "RUBY MOUNTAINS/E HUMBOLDT RANGE",
                                              "EASTERN COLUMBIA RIVER GORGE",
                                              "SOUTHERN QUEENS",
                                              "SOUTHWESTERN SAN JUAN MOUNTAINS",
                                              "PICKENS MOUNTAINS",
                                              "YORK",
                                              "WASATCH MOUNTAINS I-80 NORTH",
                                              "SOUTHERN ROCKY MOUNTAIN FRONT",
                                              "UPPER SNAKE RIVER PLAIN",
                                              "JEMEZ MOUNTAINS",
                                              "BAKERSFIELD",
                                              "GARRETT",
                                              "LUBBOCK",
                                              "BRAXTON",
                                              "BOX BUTTE"
                                  ), # National park, so should not match (FIPS = NA)
                                  state = c("COLORADO",
                                            "WYOMING",
                                            "MONTANA",
                                            "NEVADA",
                                            "OREGON",
                                            "NEW YORK",
                                            "COLORADO",
                                            "SOUTH CAROLINA",
                                            "VIRGINIA",
                                            "UTAH",
                                            "MONTANA",
                                            "IDAHO",
                                            "NEW MEXICO",
                                            "CALIFORNIA",
                                            "MARYLAND",
                                            "TEXAS",
                                            "WEST VIRGINIA",
                                            "NEBRASKA"
                                  ),
                                  state_fips = c(8,
                                                 56,
                                                 30,
                                                 32,
                                                 41,
                                                 36,
                                                 8,
                                                 45,
                                                 51,
                                                 49,
                                                 30,
                                                 16,
                                                 35,
                                                 6,
                                                 24,
                                                 48,
                                                 54,
                                                 31
                                  ),
                                  cz_fips = c(17,
                                              110,
                                              68,
                                              34,
                                              41,
                                              178,
                                              19,
                                              2,
                                              91,
                                              7,
                                              48,
                                              20,
                                              511,
                                              314,
                                              1,
                                              35,
                                              28,
                                              3
                                  )
)

match_forecast_county(counties_to_parse)
}
